refactor(renderer): let the view measure its own layout
