+++ src/modules/ui/dom.ts
@@ -37,20 +37,14 @@ import { t } from "@core/i18n";
 import { disconnectResizeObserver } from "@modules/lyrics/injectLyrics";
 import type { ThumbnailElement } from "@modules/lyrics/requestSniffer/NextResponse";
 import { getArtworkMetadata } from "@modules/lyrics/requestSniffer/requestSniffer";
-import {
-  animEngineState,
-  reflow,
-  resetAnimEngineState,
-  SCROLL_POS_OFFSET_RATIO,
-  toMs,
-} from "@modules/ui/animationEngine";
+import { animEngineState, reflow, resetAnimEngineState, toMs } from "@modules/ui/animationEngine";
 import { lyricsElementAdded } from "@modules/ui/mainLyricsView";
 import { getResumeScrollElement } from "@modules/ui/resumeScrollButton";
 import { getRequest, setRequest } from "@modules/unison/lyricsRequestTracker";
 import { getTrustTier } from "@modules/unison/trustTier";
 import type { UnisonLyricsRequest } from "@modules/unison/types";
 import { requestLyrics } from "@modules/unison/unisonApi";
-import { getRelativeLayoutBounds, log } from "@utils";
+import { log } from "@utils";
 import { generatePetName } from "@/core/keyIdentity";
 import { byId, deleteVote, type UnisonData, vote } from "../lyrics/providers/unison";
 import { buildControlsSegment, closeSourceMenu } from "./lyricsDock/controls";
@@ -1587,37 +1581,4 @@ function observeFooterForRecalc(footer: HTMLElement): void {
     lyricsElementAdded();
   });
   footerResizeObserver.observe(footer);
-}
-
-export function setExtraHeight() {
-  const lyricsElement = document.getElementsByClassName(LYRICS_CLASS)[0] as HTMLElement | undefined;
-  const tabRenderer = document.querySelector(TAB_RENDERER_SELECTOR) as HTMLElement | null;
-  if (!lyricsElement || !tabRenderer) return;
-
-  const tabRendererHeight = tabRenderer.getBoundingClientRect().height;
-  const scrollPosOffsetRatio = SCROLL_POS_OFFSET_RATIO.getNumberValue();
-  const currentPaddingBottom = Number.parseFloat(window.getComputedStyle(lyricsElement).paddingBottom) || 0;
-  const lyricsHeightWithoutBottomPadding = Math.max(0, lyricsElement.scrollHeight - currentPaddingBottom);
-
-  const lyricLines = lyricsElement.querySelectorAll<HTMLElement>(`:scope > .${LINE_CLASS}`);
-  const firstLyric = lyricLines[0] ?? null;
-  const lastLyric = lyricLines[lyricLines.length - 1] ?? null;
-  const firstLyricHeight = firstLyric ? getRelativeLayoutBounds(lyricsElement, firstLyric).height : 0;
-  const lastLyricBounds = lastLyric ? getRelativeLayoutBounds(lyricsElement, lastLyric) : null;
-
-  const paddingTop = Math.max(0, tabRendererHeight * scrollPosOffsetRatio - firstLyricHeight / 2);
-
-  document.documentElement.style.setProperty("--blyrics-padding-top", paddingTop + "px");
-
-  const lastLyricTargetContentHeight = lastLyricBounds
-    ? lastLyricBounds.y + lastLyricBounds.height / 2 + tabRendererHeight * (1 - scrollPosOffsetRatio)
-    : tabRendererHeight;
-
-  const extraHeight = Math.max(
-    lastLyricTargetContentHeight - lyricsHeightWithoutBottomPadding,
-    tabRendererHeight - lyricsHeightWithoutBottomPadding,
-    0
-  );
-
-  document.documentElement.style.setProperty("--blyrics-padding-bottom", Math.ceil(extraHeight) + "px");
 }
